Work around RSS::Parser bug
When the body of the HTTP response is empty, RSS::Parser will still not complain; the result will just be nil. With this commit, we check for that situation and raise an error ourselves. The check script will just return an empty list, as specified in http://concourse.ci/implementing-resources.html#resource-check:

> The list may be empty, if there are no versions available at the source.