Allow items with missing elements
The RSS Specification[1] states that "All elements of an item are
optional, however at least one of title or description must be present".
It is therefore incorrect to error if any of the items' elements are
omitted.

Concourse, however, requires something unique to identify the specific
resource version by, so pubDate is not allowed to not be set. This
should be work-around-able, but the specific case I'm fixing does not
require this.

[1]https://validator.w3.org/feed/docs/rss2.html#hrelementsOfLtitemgt