Last build date is not always last item date
Usually, RSS feeds are static and the date the file was last built is
the same as the publishing / update date of the most recent item,
however this is not always the case - sometimes feeds are generated
dynamically, and the build date of the file is set to `Time.now`.

In the case of a new resource being created, the pubDate is initially
set to the pubDate of the latest item.